🚚: move generator related into ./generator
