+++ codegen/generator/undgen/internal/tests/_generate_test/main.go
@@ -1,4 +1,39 @@
-package tests
+package main
+
+import (
+	"flag"
+	"fmt"
+	"os"
+	"slices"
+	"strings"
+)
+
+var (
+	excludes = flag.String("e", "", "")
+)
+
+func main() {
+	flag.Parse()
+
+	dirents, err := os.ReadDir("../testtargets")
+	if err != nil {
+		panic(err)
+	}
+
+	var names []string
+	for _, dirent := range dirents {
+		name := dirent.Name()
+		if !dirent.IsDir() || slices.Contains(strings.Split(*excludes, ","), name) {
+			continue
+		}
+		names = append(names, name)
+		f, err := os.Create(name + "_generate_test.go")
+		if err != nil {
+			panic(err)
+		}
+		_, err = fmt.Fprintf(
+			f,
+			`package tests
 
 import (
 	"maps"
@@ -6,12 +41,12 @@ import (
 	"testing"
 
 	"github.com/ngicks/go-codegen/codegen/suffixwriter"
-	"github.com/ngicks/go-codegen/codegen/undgen"
+	"github.com/ngicks/go-codegen/codegen/generator/undgen"
 	"gotest.tools/v3/assert"
 )
 
-func Test_typeparam_patcher(t *testing.T) {
-	pkgs := testTargets["typeparam"]
+func Test_%[1]s_patcher(t *testing.T) {
+	pkgs := testTargets["%[1]s"]
 	testPrinter := suffixwriter.NewTestWriter(".und_patcher", suffixwriter.WithCwd("../testtargets"))
 	err := undgen.GeneratePatcher(
 		testPrinter.Writer,
@@ -24,12 +59,12 @@ func Test_typeparam_patcher(t *testing.T) {
 	results := testPrinter.Results()
 	for _, k := range slices.Sorted(maps.Keys(results)) {
 		result := results[k]
-		t.Logf("%q:\n%s", k, result)
+		t.Logf("%%q:\n%%s", k, result)
 	}
 }
 
-func Test_typeparam_validator(t *testing.T) {
-	pkgs := testTargets["typeparam"]
+func Test_%[1]s_validator(t *testing.T) {
+	pkgs := testTargets["%[1]s"]
 	testPrinter := suffixwriter.NewTestWriter(".und_validator", suffixwriter.WithCwd("../testtargets"))
 	err := undgen.GenerateValidator(
 		testPrinter.Writer,
@@ -41,12 +76,12 @@ func Test_typeparam_validator(t *testing.T) {
 	results := testPrinter.Results()
 	for _, k := range slices.Sorted(maps.Keys(results)) {
 		result := results[k]
-		t.Logf("%q:\n%s", k, result)
+		t.Logf("%%q:\n%%s", k, result)
 	}
 }
 
-func Test_typeparam_plain(t *testing.T) {
-	pkgs := testTargets["typeparam"]
+func Test_%[1]s_plain(t *testing.T) {
+	pkgs := testTargets["%[1]s"]
 	testPrinter := suffixwriter.NewTestWriter(".und_plain", suffixwriter.WithCwd("../testtargets"))
 	err := undgen.GeneratePlain(
 		testPrinter.Writer,
@@ -58,6 +93,14 @@ func Test_typeparam_plain(t *testing.T) {
 	results := testPrinter.Results()
 	for _, k := range slices.Sorted(maps.Keys(results)) {
 		result := results[k]
-		t.Logf("%q:\n%s", k, result)
+		t.Logf("%%q:\n%%s", k, result)
+	}
+}
+`,
+			name,
+		)
+		if err != nil {
+			panic(err)
+		}
 	}
 }
