+++ codegen/generator/undgen/internal/testtargets/deeplynested/nested.und_plain.go
@@ -4,7 +4,7 @@
 package deeplynested
 
 import (
-	"github.com/ngicks/go-codegen/codegen/undgen/internal/testtargets/implementor"
+	"github.com/ngicks/go-codegen/codegen/generator/undgen/internal/testtargets/implementor"
 	"github.com/ngicks/und"
 	"github.com/ngicks/und/option"
 )
